api: fixed mis-aligned the ref count size, u8 -> u16
C++ API Modifications:
* uint8_t Paint::unref() -> uint16_t Paint::ref()
* uint8_t Paint::ref() -> uint16_t Paint::unref()
* uint8_t Paint::refCnt() -> uint16_t Paint::refCnt()

C API Modifications:
* uint8_t tvg_paint_ref(...) -> uint16_t tvg_paint_ref(...)
* uint8_t tvg_paint_unref(...) -> uint16_t tvg_paint_unref(...)
* uint8_t tvg_paint_get_ref(...) -> uint16_t tvg_paint_get_ref(...)

+++ src/bindings/capi/thorvg_capi.h
@@ -729,7 +729,7 @@ TVG_API Tvg_Result tvg_paint_del(Tvg_Paint* paint);
  *
  * @since 1.0
  */
-TVG_API uint8_t tvg_paint_ref(Tvg_Paint* paint);
+TVG_API uint16_t tvg_paint_ref(Tvg_Paint* paint);
 
 
 /**
@@ -748,7 +748,7 @@ TVG_API uint8_t tvg_paint_ref(Tvg_Paint* paint);
  *
  * @since 1.0
  */
-TVG_API uint8_t tvg_paint_unref(Tvg_Paint* paint, bool free);
+TVG_API uint16_t tvg_paint_unref(Tvg_Paint* paint, bool free);
 
 
 /**
@@ -765,7 +765,7 @@ TVG_API uint8_t tvg_paint_unref(Tvg_Paint* paint, bool free);
  *
  * @since 1.0
  */
-TVG_API uint8_t tvg_paint_get_ref(const Tvg_Paint* paint);
+TVG_API uint16_t tvg_paint_get_ref(const Tvg_Paint* paint);
 
 
 /*!
